+++ src/deploy/table_account_public.sql
@@ -5,7 +5,7 @@ CREATE TABLE vibetype.account (
 
   description TEXT CHECK (char_length(description) < 1000),
   imprint     TEXT CHECK (char_length(imprint) < 10000),
-  username    TEXT NOT NULL CHECK (char_length(username) < 100 AND username ~ '^[-A-Za-z0-9]+$') UNIQUE
+  username    TEXT NOT NULL COLLATE unicode CHECK (char_length(username) < 100 AND username ~ '^[-A-Za-z0-9]+$') UNIQUE
 );
 
 COMMENT ON TABLE vibetype.account IS E'@omit create,delete\nPublic account data.';
@@ -16,21 +16,5 @@ COMMENT ON COLUMN vibetype.account.username IS E'@omit update\nThe account''s us
 
 CREATE INDEX idx_account_username_like ON vibetype.account USING gin(username gin_trgm_ops);
 COMMENT ON INDEX vibetype.idx_account_username_like IS 'Index useful for trigram matching as in LIKE/ILIKE conditions on username.';
-
-GRANT SELECT ON TABLE vibetype.account TO vibetype_account, vibetype_anonymous;
-GRANT UPDATE ON TABLE vibetype.account TO vibetype_account;
-
-ALTER TABLE vibetype.account ENABLE ROW LEVEL SECURITY;
-
--- Make all accounts accessible by everyone.
-CREATE POLICY account_select ON vibetype.account FOR SELECT
-USING (
-  TRUE
-);
-
-CREATE POLICY account_update ON vibetype.account FOR UPDATE
-USING (
-  id = vibetype.invoker_account_id()
-);
 
 COMMIT;